+++ e2e/wallets/intergration-coins/src/index.ts
@@ -35,8 +35,8 @@ import {
     // @ts-ignore
 } from '@pioneer-platform/pioneer-coins';
 //let spec = process.env['VITE_PIONEER_URL_SPEC'] || 'https://pioneers.dev/spec/swagger.json'
-// let spec = 'https://pioneers.dev/spec/swagger.json'
-let spec = 'http://127.0.0.1:9001/spec/swagger.json'
+let spec = 'https://pioneers.dev/spec/swagger.json'
+// let spec = 'http://127.0.0.1:9001/spec/swagger.json'
 // const DB = require('@coinmasters/pioneer-db-sql');
 console.log("spec: ",spec)