+++ e2e/transfers/e2e-transfer-suite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @coinmasters/e2e-transfer-dogecoin
 
+## 1.8.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.11
+
 ## 1.8.10
 
 ### Patch Changes