+++ e2e/wallets/bip85-subaccounts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @coinmasters/integration-coins
 
+## 1.8.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.11
+
 ## 1.8.10
 
 ### Patch Changes