+++ packages/pioneer/pioneer-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @coinmasters/pioneer-sdk
 
+## 4.8.11
+
+### Patch Changes
+
+- bump
+
 ## 4.8.10
 
 ### Patch Changes